Folder restructuring, renaming according to naming conventions, reruns + debugs
